Refactoring nazw
