+++ Ankieta/src/test/resources/_readme.txt
@@ -1,0 +1,7 @@
+Milosz:
+
+Z jakiegoś powodu Unit testy  wymagają osobnego katalogu /resources
+bo nie widzą plików z głównego /resources projektu..
+co ciekawe z poziomu klasy z metodą main (tam gdzie jest "hello word") dostęp do /resources jest.
+
+proponuję zatem osobny katalog /resources dla UNIT testów, no chyba, że ktoś ma inną propozycję/rozwiązanie...
