+++ Ankieta/src/test/resources/test.json
@@ -1,0 +1,1 @@
+{"pytania":[{"id":123,"pytanie":"Jak się nazywasz?", "odpowiedzi": [{"id":1, "tresc":"nie wiem","czy_prawdziwa":false}, {"id":2, "tresc":"nie wiem", "czy_prawdziwa":false}]}]}
